+++ lib/strings.dart
@@ -26,11 +26,10 @@ bool isNotEmpty(String s) => s != null && s.isNotEmpty;
 
 /// Returns a string with characters from the given [s] in reverse order.
 ///
-/// DEPRECATED: without full support for unicode composed character sequences,
+/// NOTE: without full support for unicode composed character sequences,
 /// sequences including zero-width joiners, etc. this function is unsafe to
 /// use. No replacement is provided.
-@deprecated
-String reverse(String s) {
+String _reverse(String s) {
   if (s == null || s == '') return s;
   StringBuffer sb = new StringBuffer();
   var runes = s.runes.iterator..reset(s.length);
@@ -65,7 +64,8 @@ String loop(String s, int from, [int to]) {
     throw new ArgumentError('Input string cannot be null or empty');
   }
   if (to != null && to < from) {
-    return loop(reverse(s), -from, -to);
+    // TODO(cbracken): throw ArgumentError in this case.
+    return loop(_reverse(s), -from, -to);
   }
   int len = s.length;
   int leftFrag = from >= 0 ? from ~/ len : ((from - len) ~/ len);